gdm looping anj
