+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <uses-permission android:name="android.permission.CAMERA" />
+
     <application
+        android:requestLegacyExternalStorage="true"
         android:label="EverWrite"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
