+++ pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   image_picker: ^1.0.4
+  dotted_border: ^2.1.0
   logger: ^2.0.2+1
 
 dev_dependencies:
